Move bag_of_words into root dir
